[cli][core] Move with_klio_config from cli to core to allow useage in klioexec
