+++ cli/src/klio_cli/options.py
@@ -15,7 +15,7 @@
 
 import click
 
-import klio_core.options as core_options
+from klio_core import options as core_options
 
 
 def _verify_gcs_uri(ctx, param, value):
@@ -31,75 +31,6 @@ def _verify_gcs_uri(ctx, param, value):
 #####
 # common options
 #####
-def job_dir(*args, **kwargs):
-    mutually_exclusive = kwargs.get("mutex", [])
-
-    def wrapper(func):
-        return click.option(
-            "-j",
-            "--job-dir",
-            type=click.Path(exists=True),
-            help=(
-                "Job directory where the job's ``Dockerfile`` is located. "
-                "Defaults current working directory."
-            ),
-            cls=core_options.MutuallyExclusiveOption,
-            mutually_exclusive=mutually_exclusive,
-        )(func)
-
-    # allows @options.foo to be used without parens (i.e. no need to do
-    # `@options.foo()`) when there are no args/kwargs provided
-    if args:
-        return wrapper(args[0])
-    return wrapper
-
-
-def config_file(*args, **kwargs):
-    mutually_exclusive = kwargs.get("mutex", [])
-
-    def wrapper(func):
-        return click.option(
-            "-c",
-            "--config-file",
-            type=click.Path(exists=False),
-            help=(
-                "Path to config filename. If ``PATH`` is not absolute, it "
-                "will be treated relative to ``--job-dir``. Defaults to "
-                "``klio-job.yaml``."
-            ),
-            cls=core_options.MutuallyExclusiveOption,
-            mutually_exclusive=mutually_exclusive,
-        )(func)
-
-    # allows @options.foo to be used without parens (i.e. no need to do
-    # `@options.foo()`) when there are no args/kwargs provided
-    if args:
-        return wrapper(args[0])
-    return wrapper
-
-
-def override(func):
-    return click.option(
-        "-O",
-        "--override",
-        default=[],
-        multiple=True,
-        help="Override a config value, in the form ``key=value``.",
-    )(func)
-
-
-def template(func):
-    return click.option(
-        "-T",
-        "--template",
-        default=[],
-        multiple=True,
-        help=(
-            "Set the value of a config template parameter"
-            ", in the form ``key=value``.  Any instance of ``${key}`` "
-            "in ``klio-job.yaml`` will be replaced with ``value``."
-        ),
-    )(func)
 
 
 def force_build(func):
