+++ cli/src/klio_cli/utils/config_utils.py
@@ -16,24 +16,9 @@
 import logging
 import warnings
 
-import yaml
-
-
 SUPPORTED_CONFIG_VERSIONS = (2,)
 DEPRECATED_CONFIG_VERSIONS = (1,)
 ALL_CONFIG_VERSIONS = SUPPORTED_CONFIG_VERSIONS + DEPRECATED_CONFIG_VERSIONS
-
-
-def get_config_by_path(config_filepath, parse_yaml=True):
-    try:
-        with open(config_filepath) as f:
-            if parse_yaml:
-                return yaml.safe_load(f)
-            else:
-                return f.read()
-    except IOError:
-        logging.error("Could not read config file {0}".format(config_filepath))
-        raise SystemExit(1)
 
 
 # TODO: integrate this into KlioConfig as a converter
