+++ core/setup.py
@@ -165,6 +165,7 @@ EXTRAS_REQUIRE = {
     "docs": ["sphinx", "interrogate"],
     "tests": [
         "coverage",
+        "mock",  # for py2 - remove when dropping support
         "pytest>=4.3.0",  # 4.3.0 dropped last use of `convert`
         "pytest-cov",
         "pytest-mock",
