+++ core/tests/test_utils.py
@@ -13,6 +13,10 @@
 # limitations under the License.
 #
 
+import os
+
+from unittest import mock
+
 import pytest
 
 from google.api_core import exceptions as gapi_exceptions
@@ -119,3 +123,114 @@ def test_get_publisher(in_globals, mock_publisher):
 
     assert client == ret_publisher
     delattr(utils, "klio_global_state_publisher_a-topic")
+
+
+#########################
+# Config Utils tests
+#########################
+@pytest.fixture
+def patch_os_getcwd(monkeypatch, tmpdir):
+    test_dir = str(tmpdir.mkdir("testing"))
+    monkeypatch.setattr(os, "getcwd", lambda: test_dir)
+    return test_dir
+
+
+def test_get_config_by_path(mocker, monkeypatch):
+    m_open = mocker.mock_open()
+    mock_open = mocker.patch("klio_core.utils.open", m_open)
+
+    mock_safe_load = mocker.Mock()
+    monkeypatch.setattr(utils.yaml, "safe_load", mock_safe_load)
+
+    path = "path/to/a/file"
+    utils.get_config_by_path(path)
+
+    mock_open.assert_called_once_with(path)
+
+
+def test_get_config_by_path_error(mocker, monkeypatch, caplog):
+    m_open = mocker.mock_open()
+    mock_open = mocker.patch("klio_core.utils.open", m_open)
+    mock_open.side_effect = IOError
+
+    with pytest.raises(SystemExit):
+        path = "path/to/a/file"
+        utils.get_config_by_path(path)
+
+    assert 1 == len(caplog.records)
+
+
+#########################
+# Cli/exec command tests
+#########################
+@pytest.mark.parametrize(
+    "image",
+    (
+        "dataflow.gcr.io/v1beta3/python",
+        "dataflow.gcr.io/v1beta3/python-base",
+        "dataflow.gcr.io/v1beta3/python-fnapi",
+    ),
+)
+def test_warn_if_py2_job(image, patch_os_getcwd, mocker):
+    dockerfile = (
+        '## -*- docker-image-name: "gcr.io/foo/bar" -*-\n'
+        "FROM {image}:1.2.3\n"
+        'LABEL maintainer "me@example.com"\n'
+    ).format(image=image)
+
+    m_open = mock.mock_open(read_data=dockerfile)
+    mock_open = mocker.patch("klio_core.utils.open", m_open)
+
+    warn_msg = (
+        "Python 2 support in Klio is deprecated. "
+        "Please upgrade to Python 3.5+"
+    )
+    with pytest.warns(UserWarning, match=warn_msg):
+        utils.warn_if_py2_job(patch_os_getcwd)
+
+    exp_read_file = os.path.join(patch_os_getcwd, "Dockerfile")
+    mock_open.assert_called_once_with(exp_read_file, "r")
+
+
+@pytest.mark.parametrize("has_from_line", (True, False))
+def test_warn_if_py2_job_no_warn(has_from_line, patch_os_getcwd, mocker):
+    from_line = "\n"
+    if has_from_line:
+        from_line = "FROM dataflow.gcr.io/v1beta3/python36-fnapi:1.2.3\n"
+
+    dockerfile = (
+        '## -*- docker-image-name: "gcr.io/foo/bar" -*-\n'
+        + from_line
+        + 'LABEL maintainer "me@example.com"\n'
+    )
+
+    m_open = mock.mock_open(read_data=dockerfile)
+    mock_open = mocker.patch("klio_core.utils.open", m_open)
+
+    utils.warn_if_py2_job(patch_os_getcwd)
+
+    exp_read_file = os.path.join(patch_os_getcwd, "Dockerfile")
+    mock_open.assert_called_once_with(exp_read_file, "r")
+
+
+@pytest.mark.parametrize(
+    "job_dir,conf_file",
+    (
+        (None, None),
+        (None, "klio-job2.yaml"),
+        ("foo/bar", None),
+        ("foo/bar", "klio-job2.yaml"),
+    ),
+)
+def test_get_config_job_dir(job_dir, conf_file, patch_os_getcwd):
+    exp_job_dir = patch_os_getcwd
+    if job_dir:
+        exp_job_dir = os.path.abspath(os.path.join(patch_os_getcwd, job_dir))
+    exp_conf_file = conf_file or os.path.join(exp_job_dir, "klio-job.yaml")
+    if job_dir and conf_file:
+        exp_conf_file = os.path.join(job_dir, conf_file)
+
+    ret_job_dir, ret_conf_file = utils.get_config_job_dir(job_dir, conf_file)
+
+    assert exp_job_dir == ret_job_dir
+    assert exp_conf_file == ret_conf_file
