+++ docs/src/reference/core/api/options.rst
@@ -1,5 +1,5 @@
 ``klio_core.options``
-===================
+=====================
 
 .. automodule:: klio_core.options
