+++ docs/src/reference/core/api/utils.rst
@@ -10,3 +10,6 @@
 .. autofunction:: get_global
 .. autofunction:: set_global
 .. autofunction:: delete_global
+.. autofunction:: get_config_by_path
+.. autofunction:: get_config_job_dir
+.. autofunction:: with_klio_config
