+++ exec/src/klio_exec/cli.py
@@ -22,7 +22,7 @@ import click
 import yaml
 
 from klio_core import config
-import klio_core.options as core_options
+from klio_core import options as core_options
 
 from klio_exec import options
 from klio_exec.commands import audit
